requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==4.0.0
distlib==0.3.1
etcd3==0.12.0
filelock==3.0.12
gevent==1.4.0
google==3.0.0
greenlet==1.0.0
grpcio==1.36.1
gsocketpool==0.1.6
idna==2.10
idna-ssl==1.1.0
importlib-metadata==3.7.3
importlib-resources==5.1.2
multidict==5.1.0
pipenv==2020.11.15
protobuf==3.15.6
py-bee-config==0.1.0
py-bee-util==0.1.1
py-consul==1.2.2
PyYAML==5.4.1
requests==2.25.1
requests-unixsocket==0.2.0
six==1.15.0
soupsieve==2.2
tenacity==7.0.0
typing-extensions==3.7.4.3
urllib3==1.26.4
virtualenv==20.4.3
virtualenv-clone==0.5.4
yarl==1.6.3
zipp==3.4.1
